NTP-315: Share search model between search results and tuition partner
The model is shared using TempData because `asp-route-all-data` does not work nicely when a route includes parameters - `@page {id}`.

It *is* possible to use multiple `asp-route-{parameter}` directives, however this defeats the benefit of having a `SearchModel` able to be fed into `asp-all-route-data` - namely that any additions to the model cannot be forgotten in individual `asp-page` links.