NTP-462: Randomise search results (#81)
* NTP-462: Display a stable pseudo-randomisation of the results

* NTP-462: Speed up randomness tests

* Dotnet format

* Update content to reflect results are sorted randomly

* Pre-order random results by name rather than database ID

This attempts to preserve the random order after the spreadsheets are updated.  The import process removes the existing tuition partner data, and then creates new records for each TP.  As that will change the db IDs, it is likely to alter the order the records are in before randomisation, which will result in a different random order for the same search criteria.

Ordering the results by name prior to randomisation will keep a stable ordering even after the row ID changes.