NTP-632: URLs only contain URL safe characters
Some tuition partner names have characters that do not safe for URLs. A case in point is `Eric's Tutoring`. This results in the logos not appearing correctly as the browser looks for `eric%23s-tutoring` instead of `eric's-tutoring`.

The resolution is to remove all unsafe characters, resulting in `erics-tutoring`.

The ASPNet core engine does rely on the brace and parenthesis characters remaining untouched as they are used by the routing engine, and our renaming behaviour would break routing if these characters were removed.